requirements.txt
Pillow==9.5.0
Pillow==10.0.0
PySimpleGUI==4.60.4
PySimpleGUI==4.60.5
